requirements_for_win.txt
anndata==0.10.7
array_api_compat==1.7.1
asttokens==2.4.1
attrs==23.2.0
blinker==1.8.2
cached-property==1.5.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
ConfigArgParse==1.7
contourpy==1.2.1
cycler==0.12.1
dash==2.17.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.8.1
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.1
Flask==3.0.3
fonttools==4.53.0
future==1.0.0
get-annotations==0.1.2
h5py==3.11.0
idna==3.7
igraph==0.11.5
importlib_metadata==7.1.0
ipykernel==6.29.4
ipython==8.25.0
ipywidgets==8.1.3
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_widgets==3.0.11
kiwisolver==1.4.5
legacy-api-wrap==1.4
leidenalg==0.10.2
llvmlite==0.42.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
munkres==1.1.4
natsort==8.4.0
nbformat==5.10.4
nest_asyncio==1.6.0
networkx==3.3
numba==0.59.1
numpy==1.26.4
open3d==0.18.0
opencv-python==4.10.0.82
packaging==24.0
pandas==2.2.2
parso==0.8.4
patsy==0.5.6
pickleshare==0.7.5
pillow==10.3.0
pip==24.0
platformdirs==4.2.2
plotly==5.22.0
prompt_toolkit==3.0.46
psutil==5.9.8
pure-eval==0.2.2
pycpd==2.0.0
Pygments==2.18.0
pynndescent==0.5.12
pyparsing==3.1.2
python-dateutil==2.9.0
pytz==2024.1
pywin32==306
pyzmq==26.0.3
referencing==0.35.1
requests==2.32.3
retrying==1.3.4
rpds-py==0.18.1
scanpy==1.10.1
scikit-learn==1.5.0
scipy==1.13.1
seaborn==0.13.2
session-info==1.0.0
setuptools==70.0.0
six==1.16.0
stack-data==0.6.2
statsmodels==0.14.2
stdlib-list==0.10.0
tenacity==8.3.0
texttable==1.7.0
threadpoolctl==3.5.0
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
umap-learn==0.5.5
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.3
wheel==0.43.0
widgetsnbextension==4.0.11
zipp==3.17.0
